% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aging.R
\name{aging}
\alias{aging}
\title{Module for aging over time}
\usage{
aging(dat, at)
}
\arguments{
\item{dat}{object with patients' data}
}
\value{
This function returns \code{dat} after updating ' \code{age}.
}
\description{
Module for aging over time
}
