% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrival.R
\name{arrival}
\alias{arrival}
\title{arrival records the newly-infected individuals at the previous time period, to add new individuals to the cohort}
\usage{
arrival(dat, at, scenario)
}
\arguments{
\item{dat}{the current data on the cohort}

\item{at}{the current date}

\item{scenario}{scenario's name}
}
\value{
This function returns the \code{dat} object with updated treatment status for each individual, and the cohort summary
in \code{dat$epi} with the number of lost to follow-up
\code{lfu}, and the number of patients treated \code{onTrt} at each time unit.
}
\description{
arrival records the newly-infected individuals at the previous time period, to add new individuals to the cohort
}
