% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_outcomes.R
\name{calculate_DALYs}
\alias{calculate_DALYs}
\title{Calculate Disability-Adjusted-Life-Years for each scenario and each simulation.
One DALYs can be thought of as one healthy year of life lost.}
\usage{
calculate_DALYs(res_sim, parameters_bio, param)
}
\arguments{
\item{res_sim}{the main health outcomes of interest for different simulation and a given scenario}
}
\value{
`DALYs` a list object with the number of DALYs for each simulation
}
\description{
\code{calculate_DALYs} transforms the number of deaths, new infections,
patients on and off treatment into a summary metric.
}
\details{
This function uses disability weights from the Global Burden of Disease Study to summarize
the number of DALYs in each of the scenario and therefore deduced the DALYs averted thanks
to the intervention.
}
