% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd4.R
\name{cd4}
\alias{cd4}
\title{cd4 module}
\usage{
cd4(dat, at)
}
\arguments{
\item{dat}{data}

\item{at}{parameter}
}
\value{
This function returns \code{dat} after updating individuals'
\code{cd4Count}.
}
\description{
Module for assigning to newly HIV-infected individuals their baseline CD4 count (i.e. CD4 count at time of infection) 
             and updating individuals cd4 count at each time step, depending on individuals' treatment status, age and sex
}
