% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{compare_table}
\alias{compare_table}
\title{Return comparison table}
\usage{
compare_table(data, scen_1, scen_2)
}
\arguments{
\item{summary_results}{object with simulations results (deaths, new infections, # patients off and on treatment, # of lost to follow-up)
from both the baseline and intervention scenarios}
}
\value{

}
\description{
Return comparison table
}
