% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{create_scenario_list}
\alias{create_scenario_list}
\title{Parameters for agent-based model simulation}
\usage{
create_scenario_list(scenarios_folder = "params/scenarios/")
}
\arguments{
\item{size_active_line}{size observed active line of HIV patients - used to deduce the size of the catchment area}

\item{age.mean}{mean age cohort patients}

\item{age.median}{median age cohort patients}

\item{base.cd4}{cd4 count at infection}

\item{di.cd4.aids}{cd4 count AIDS}

\item{di.cd4.rate}{daily death rate AIDS}

\item{base.tprob.male}{baseline probability transmission male}

\item{base.tprob.feml}{baseline probability transmission female}

\item{act.rate.early}{daily rate sexual act in early phases of infection}

\item{act.rate.late}{daily rate sexual act in later stages of infection}

\item{act.rate.cd4}{threshold cd4 count value at which patients 
switch from early to later stages}

\item{tx.cd4.recrat.feml}{daily increase in cd4 count after treatment initiation for females"}

\item{tx.cd4.recrat.male}{daily increase in cd4 count after treatment initiation for males}

\item{tx.cd4.decrat.feml}{daily decrease in cd4 count after treatment interruption for females}

\item{tx.cd4.decrat.male}{daily decrease in cd4 count after treatment interruption for males}

\item{tx.vlsupp.time}{time in weeks from treatment initiation to viral suppression}

\item{tx.vlsupp.level}{}

\item{vl.acute.topeak}{time in weeks to peak viremia during acute infection}

\item{vl.acute.toset}{time in weeks to viral set point following peak viremia}

\item{vl.acute.peak}{Log 10 viral load at acute peak}

\item{vl.acute.setpoint}{Log 10 viral load at set point}

\item{vl.aidsmax}{Maximum log 10 viral load during AIDS}

\item{hiv.coverage}{National annual testing coverage}

\item{tx.cd4.detect}{Minimum cd4 count after which patients seek care at health facilities}

\item{tx.init.cd4.mean}{Mean cd4 count at ART initiation}

\item{tx.init.cd4.sd}{Standard deviation cd4 count at ART initiation}

\item{tx.coverage}{ART coverage}

\item{tx.adhere.full}{proportion of fully adherent patients}

\item{tx.adhere.part}{Proportion of patients, among those that are not fully adherent,
that take their treatment from time to time}

\item{lfu.prop}{Annual proportion of patients being lost to follow-up}
}
\description{
Sets the epidemic parameters for stochastic agent-based model
             simulated with \code{\link{run_abm}}.
}
\details{
: Load the different parameters used later on to simulate the different components of the agent-based model
}
\section{Biological parameters}{

Parameters that are supposed to remain the same across the different scenarios, i.e. parameters
over which the intervention has no impact. Parameters' values were found in the literature.
They can be modified based on new evidence, or measurements.
}

\section{Intervention parameters}{

Parameters that are impacted by the simulation, and whose value changes from one scenario
to the other.
}

\seealso{
Use @param time.unit any time step [default = 7], and @param list_scenario
         specified in a `yaml`. Run the parameterized model with \code{\link{run_abm}}
}
