% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{DALY_comparison}
\alias{DALY_comparison}
\title{Making Daly Comparison plots}
\usage{
DALY_comparison(daly)
}
\arguments{
\item{daly}{object with simulations results in terms of DALYs from both the baseline and intervention scenarios}
}
\value{

}
\description{
Making Daly Comparison plots
}
