% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departure.R
\name{deaths}
\alias{deaths}
\title{: departure module}
\usage{
deaths(dat, at)
}
\arguments{
\item{dat}{}

\item{at}{}
}
\value{
This function returns the \code{dat} object recording for each time step \code{at}, the number of susceptible deaths 
\code{ds.flow}, and infected deaths \code{di.flow}
broken down into on treatment and lost to follow-up deaths \code{dtrt.flow} and \code{dlfu.flow} (therefore, `di.flow = dtrt.flow + dlfu.flow`)
}
\description{
: Stochastically determined deaths at each time period of the simulation
}
\seealso{
This function is used by:
  \code{\link{run_abm}}
}
