% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deleteAttr}
\alias{deleteAttr}
\title{Delete the attribute of an individual
\code{deleteAttr} delete the attribute of an individual that exits the simulation.}
\usage{
deleteAttr(attrList, ids)
}
\arguments{
\item{attrList}{list of attributes of individuals in the simulation}

\item{ids}{Numeric vector identifying the individuals to delete from the simulation}
}
\value{
attrList, list of attributes of individuals remaining in the simulation
}
\description{
Delete the attribute of an individual
\code{deleteAttr} delete the attribute of an individual that exits the simulation.
}
\details{
In the default version of the simulation,
individuals exit the simulation if `active == 0`, i.e. if they die.
}
