% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadEstimates.R
\name{extract_unaids}
\alias{extract_unaids}
\title{Extract data from a given pays from the UNAIDS database}
\usage{
extract_unaids(
  python_script,
  country,
  url = "https://aidsinfo.unaids.org/",
  outpath
)
}
\arguments{
\item{python_script}{path to unaids_scrap.py}

\item{country}{country of interest}

\item{url}{url to unaids website}

\item{outpath}{path to the file where the extraction will be written}
}
\description{
Extract data from a given pays from the UNAIDS database
}
