% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{init}
\alias{init}
\title{initialize module}
\usage{
init(i.prev.male = 1, i.prev.feml = 1, max.inf.time = 5 * 365, n = 1000, ...)
}
\arguments{
\item{i.prev.male}{prevalence of the infection in males
[default = 1]}

\item{i.prev.feml}{prevalence of the infection in females
[default = 1]}

\item{max.inf.time}{oldest infection considered in the simulation}

\item{n}{number of patients in the simulated cohort}
}
\value{
: This function returns
         (1) the \code{p} object with additional parameters
         (2) the \code{dat} object with initial \code{male}, \code{active},\code{facility}, \code{age}, and \code{entTime} attributes
         (3) the \code{dat} object with initial \code{status}
         (4) the \code{dat} object with initial \code{ageInf}
         (5) the \code{dat} object with initial \code{dxStatus} and \code{dxTime}
         (6) the \code{dat} object with initial \code{txCD4min}, \code{txCD4start}, and \code{txType}
}
\description{
These scripts contain the functions that set the initial conditions for the models.
}
