% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadEstimates.R
\name{load_estimates}
\alias{load_estimates}
\title{Load extracts from UNAIDS' website to usable estimates}
\usage{
load_estimates(extracted_data, output)
}
\arguments{
\item{extracted_data}{adress of the extracted datga}

\item{output}{adress where the result of the function will be written}
}
\value{
a rds file of the usable parameters
}
\description{
Load extracts from UNAIDS' website to usable estimates
}
