% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_pyramid}
\alias{make_pyramid}
\title{Making Age Pyramid}
\usage{
make_pyramid(summary_results)
}
\arguments{
\item{summary_results}{object with simulations results (deaths, new infections, # patients off and on treatment, # of lost to follow-up)
from both the baseline and intervention scenarios}
}
\value{

}
\description{
Making Age Pyramid
}
