% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{prevalence}
\alias{prevalence}
\title{Prevalence module}
\usage{
prevalence(dat, at)
}
\arguments{
\item{dat}{}

\item{at}{}
}
\value{
This function returns the \code{dat} object with updated epidemiological quantities for each
time step.
}
\description{
Module function for keeping track of epidemiological quantities of interest
             including the number of virally suppressed agents `vlSupp`, 
             adherence over time.
}
\seealso{
This function is used by:
  \code{\link{run_abm}}
}
