% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proba_to_rate}
\alias{proba_to_rate}
\title{\code{proba_to_rate} transforms annual probability of a given event
into a rate of occurrence of this event for the specified time unit \code{t}}
\usage{
proba_to_rate(p, t = 365.25)
}
\arguments{
\item{p}{the annual probability (must be between 0 and 1)}

\item{t}{the unit of time in which the probability was expressed (by default, a year)}
}
\value{
`1-exp(-r)` the rate of occurence of the event per \code{time.unit}
}
\description{
\code{proba_to_rate} transforms annual probability of a given event
into a rate of occurrence of this event for the specified time unit \code{t}
}
\details{
This function assumes a constant rate of occurrence of an event over a given time period
}
