% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{result_comparison_plot}
\alias{result_comparison_plot}
\title{Module to produce graphical outputs to visualize simulations' results}
\usage{
result_comparison_plot(res)
}
\arguments{
\item{res}{object with simulations results (deaths, new infections, # patients off and on treatment, # of lost to follow-up)
from both the baseline and intervention scenarios}

\item{daly}{object with simulations results in terms of DALYs from both the baseline and intervention scenarios}
}
\value{
This function returns summary plots of the simulations' results.
}
\description{
Module to produce graphical outputs to visualize simulations' results
}
