% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_abm.R
\name{run_abm}
\alias{run_abm}
\title{run_abm.R}
\usage{
run_abm(
  init,
  param,
  scenario_simulation,
  intervention_start = 0,
  prop_male,
  nsteps = 52 * 5
)
}
\arguments{
\item{init}{A list including the initial conditions of the simulation:
prevalence of the infection, maximum span of infection considered, initial size of the cohort simulated}

\item{param}{A list including the biological and intervention parameters, along with the age-sex-specific death rates}

\item{scenario_simulation}{Character vector indicating which scenario is being considered for the simulation}

\item{intervention_start}{A numeric value indicating at which time step the intervention starts
(i.e. when the simulation starts using the intervention parameters over the baseline parameters)
[default = 0]}

\item{prop_male}{A numeric value indicating what share of the population is male
This value is currently scraped from the UNAIDS indicators}

\item{n_steps}{A numeric value indicating the duration of the intervention
[default = 52*5], i.e. 5 years}
}
\value{
This function returns the \code{dat} object, recording the evolution of the attributes and epidemiological modules,
        for the duration of the simulation
}
\description{
Run the different functions to get the full simulation of a cohort
of patients, in the baseline scenario (standard of care) and with the different intervention scenarios
}
\seealso{
This function is used by:
  \code{\link{run_simulations}}
}
