% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSimulation.R
\name{run_simulations}
\alias{run_simulations}
\title{run_simulation.R}
\usage{
run_simulations(
  init,
  param,
  scenario_simulation,
  intervention_start = 0,
  prop_male,
  nsteps = 52 * 5,
  nsim = 100
)
}
\arguments{
\item{init}{A list including the initial conditions of the simulation:
prevalence of the infection, maximum span of infection considered, initial size of the cohort simulated}

\item{param}{A list including the biological and intervention parameters, along with the age-sex-specific death rates}

\item{scenario_simulation}{Character vector indicating which scenario is being considered for the simulation}

\item{intervention_start}{A numeric value indicating at which time step the intervention starts
(i.e. when the simulation starts using the intervention parameters over the baseline parameters)
[default = 0]}

\item{prop_male}{A numeric value indicating what share of the population is male
This value is currently scraped from the UNAIDS indicators}

\item{nsim}{[default = 1000]}

\item{n_steps}{A numeric value indicating the duration of the intervention
[default = 52*5], i.e. 5 years}
}
\value{
This function returns the \code{out} collecting the summary outcomes of interest
}
\description{
Run agent-based models `nsim` times
under the different scenarios provided by the user
}
