% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrival.R
\name{setNewAttr}
\alias{setNewAttr}
\title{Assign attributes to the newly "created" individuals in the cohort}
\usage{
setNewAttr(dat, at, nNew, scenario)
}
\arguments{
\item{dat}{the current cohort}

\item{at}{the current date}

\item{nNew}{the number of new patients}

\item{scenario}{the scenario's name}
}
\value{
This function returns the \code{dat} object with updated treatment status for each individual, and the cohort summary
in \code{dat$epi} with the number of lost to follow-up
\code{lfu}, and the number of patients treated \code{onTrt} at each time unit.
}
\description{
Assign attributes to the newly "created" individuals in the cohort
}
