% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_outcomes.R
\name{summary_outcomes}
\alias{summary_outcomes}
\title{summarize the main health outcomes of the simulation}
\usage{
summary_outcomes(simul_results)
}
\arguments{
\item{simul_results}{the object storing the population metrics and simulation outcomes across every time step of the simulation}
}
\value{
A list of outputs - the main health outcomes of interest for the simulation
}
\description{
summarize the main health outcomes of the simulation
}
