% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hivtest.R
\name{test}
\alias{test}
\title{: test module}
\usage{
test(dat, at, scenario)
}
\arguments{
\item{dat}{}

\item{at}{}

\item{scenario}{scenario's name}
}
\value{
: This function returns the \code{dat} object with updated \code{dxStat}, \code{txStat}, and \code{dxTime} attributes
}
\description{
: Stochastically determined which individuals are getting tested at every time step `at'
}
\seealso{
This function is used by:
  \code{\link{run_abm}}
}
