% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{transmission}
\alias{transmission}
\title{transmission module}
\usage{
transmission(dat, at)
}
\arguments{
\item{dat}{}

\item{at}{}
}
\value{
This function returns the epidemiological module \code{dat$epi} after calculating the number of new infection
\code{si.flow} at each time step.
}
\description{
Calculate the number of new infections at each time step based on biological parameters
             (the baseline probability of transmission per act \code{base.tprob}, the number of act per time unit), 
             and the viral load level of each individual
}
\seealso{
This function is used by:
  \code{\link{run_abm}}
}
