% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hivtrt.R
\name{trt}
\alias{trt}
\title{treatment module}
\usage{
trt(dat, at, scenario)
}
\arguments{
\item{dat}{}

\item{at}{}

\item{scenario}{scenario's name}
}
\value{
This function returns the \code{dat} object with updated treatment status for each individual, and the cohort summary
in \code{dat$epi} with the number of lost to follow-up
\code{lfu}, and the number of patients treated \code{onTrt} at each time unit.
}
\description{
Module function for anti-retroviral treatment initiation and
             adherence over time.
}
\seealso{
This function is used by:
  \code{\link{run_abm}}
}
