% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl.R
\name{vl}
\alias{vl}
\title{HIV viral load module}
\usage{
vl(dat, at)
}
\arguments{
\item{dat}{}

\item{at}{}
}
\value{
This function returns \code{dat} after updating individuals'
\code{vlLevel}.
}
\description{
Module for assigning HIV-infected individuals' baseline viral load (i.e. viral load after infection)
             and updating individuals viral load at each time step, depending on individuals' treatment status, age and CD4 count
}
\seealso{
This function is used by:
  \code{\link{run_abm}}
}
